Fix bug:解决的pe_ttm的问题

+++ AnyQuant/AnyQuant_Common/src/main/java/dataservice/APIInterfaceImpl.java
@@ -218,7 +218,7 @@ public class APIInterfaceImpl implements APIInterface{
 		List<StockPO> stocks =  new  ArrayList<>();
 			for(int i=0;i<trading_info.size();i++){
 				StockPO stock  = MyJSONObject.toBean(trading_info.getJSONObject(i), StockPO.class);
-				stock.setPe_ttm(Float.parseFloat(trading_info.getJSONObject(i).getString("pe_ttm")));
+				stock.setPe_ttm(Double.parseDouble(trading_info.getJSONObject(i).getString("pe_ttm")));
 				stock.setAdj_price(Double.parseDouble(trading_info.getJSONObject(i).getString("adj_price")));
 				stock.setCode(stockCode);
 				stocks.add(stock);
